requirements.txt
django-cms<3.5
djangocms-admin-style>=1.2,<1.3
django-treebeard>=4.0,<5.0

djangocms-text-ckeditor>=3.2.1
djangocms-link>=1.8
djangocms-style>=1.7
djangocms-googlemap>=0.5
djangocms-snippet>=1.9
djangocms-video>=2.0
djangocms-column>=1.6
easy_thumbnails
django-filer>=1.2
cmsplugin-filer>=1.1
Django<1.10
pytz
django-classy-tags>=0.7
html5lib>=0.999999,<0.99999999
Pillow>=3.0
django-sekizai>=0.9
django-select2<5.0six

# heroku
gunicorn==19.6.0
dj-database-url==0.4.1
whitenoise==3.2
psycopg2
django_storages==1.4.1
boto==2.41.0

# kitchensink
django-hvad==1.6.0
django-admin-sortable2==0.6.4
django-multiselectfield==0.1.4
django-admin-tools==0.8.0
